% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_institutes.R
\name{best_institutes}
\alias{best_institutes}
\title{Jee College Predictor}
\usage{
best_institutes(
  w,
  r = 1000,
  category = "General",
  institute_type = "IIT",
  quota = "AI"
)
}
\arguments{
\item{w}{}

\item{r}{}
}
\description{
Best College according to rank
}
\examples{
best_institutes(w , rank)

}
